+++ types/quote.go
@@ -11,12 +11,12 @@ type Quote struct {
 	PenaltyFee         uint64 `json:"penaltyFee" db:"penalty_fee"`
 	ContractAddr       string `json:"contractAddr" db:"contract_addr"`
 	Data               string `json:"data" db:"data"`
-	GasLimit           uint   `json:"gasLimit" db:"gas_limit"`
+	GasLimit           uint32 `json:"gasLimit" db:"gas_limit"`
 	Nonce              int64  `json:"nonce" db:"nonce"`
 	Value              uint64 `json:"value" db:"value"`
-	AgreementTimestamp uint   `json:"agreementTimestamp" db:"agreement_timestamp"`
-	TimeForDeposit     uint   `json:"timeForDeposit" db:"time_for_deposit"`
-	CallTime           uint   `json:"callTime" db:"call_time"`
-	Confirmations      uint   `json:"confirmations" db:"confirmations"`
+	AgreementTimestamp uint32 `json:"agreementTimestamp" db:"agreement_timestamp"`
+	TimeForDeposit     uint32 `json:"timeForDeposit" db:"time_for_deposit"`
+	CallTime           uint32 `json:"callTime" db:"call_time"`
+	Confirmations      uint16 `json:"confirmations" db:"confirmations"`
 	CallOnRegister     bool   `json:"callOnRegister" db:"call_on_register"`
 }
